requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.5
django-extensions==3.1.0
django-filter==2.4.0
django-mathfilters==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
entrypoints==0.3
fire==0.4.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.4.3
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==5.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy==0.800
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==1.0.18
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.5
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
